Split store module to event-store-api and inmemory implementation.
