+++ event-store-api/pom.xml
@@ -9,7 +9,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>micro-modeling</artifactId>
+    <artifactId>event-store-api</artifactId>
 
     <properties>
         <maven.compiler.source>22</maven.compiler.source>
@@ -19,12 +19,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>io.event.thinking</groupId>
-            <artifactId>event-store-api</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
             <groupId>io.projectreactor</groupId>
             <artifactId>reactor-core</artifactId>
         </dependency>
@@ -32,12 +26,6 @@
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-engine</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>io.projectreactor</groupId>
-            <artifactId>reactor-test</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
