+++ event-store-api/src/main/java/io/event/thinking/eventstore/api/InvalidConsistencyConditionException.java
@@ -1,0 +1,15 @@
+package io.event.thinking.eventstore.api;
+
+/**
+ * The exception thrown in the case when the {@link ConsistencyCondition} of the
+ * {@link EventStore#append(Event, ConsistencyCondition)} is not met.
+ */
+public class InvalidConsistencyConditionException extends RuntimeException {
+
+    /**
+     * Creates the {@link InvalidConsistencyConditionException}.
+     */
+    public InvalidConsistencyConditionException() {
+        super();
+    }
+}
