+++ inmemory-event-store/pom.xml
@@ -9,7 +9,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>micro-modeling</artifactId>
+    <artifactId>inmemory-event-store</artifactId>
 
     <properties>
         <maven.compiler.source>22</maven.compiler.source>
